+++ Lightcore/Constants.cs
@@ -6,8 +6,8 @@ namespace Lightcore
     {
         public const int DeltaDecimals = 10;
         public static string DeltaFormat = $"n{DeltaDecimals}";
-        public static double Delta = Math.Pow(0.1f, DeltaDecimals);
-        public const double PI = Math.PI;
-        public const double PI2 = PI * 2;
+        public static float Delta = (float)Math.Pow(0.1f, DeltaDecimals);
+        public const float PI = (float)Math.PI;
+        public const float PI2 = PI * 2;
     }
 }
