+++ Lightcore/Lighting/Models/AngleLight.cs
@@ -5,12 +5,12 @@
 
     public class AngleLight : DirectLight
     {
-        public AngleLight (Vector color, Vector position, Vector direction, double strength, double angle, Guid? origin = null, Guid? id = null, int generation = 0) : base (color, position, direction, strength, origin, id, generation)
+        public AngleLight (Vector color, Vector position, Vector direction, float strength, float angle, Guid? origin = null, Guid? id = null, int generation = 0) : base (color, position, direction, strength, origin, id, generation)
         {
             Angle = angle;
         }
 
-        public double Angle { get; set; }
+        public float Angle { get; set; }
 
         public override Light Clone()
         {
