+++ Lightcore/Lighting/Models/DirectLight.cs
@@ -5,7 +5,7 @@
 
     public abstract class DirectLight : Light
     {
-        public DirectLight(Vector color, Vector position, Vector direction, double strength, Guid? origin = null, Guid? id = null, int generation = 0) 
+        public DirectLight(Vector color, Vector position, Vector direction, float strength, Guid? origin = null, Guid? id = null, int generation = 0) 
             : base (color, position, strength, origin, id, generation)
         {
             Direction = direction;
